refactor: 컨트롤러의 게임 로직을 RacingGameService로 분리
- 랜덤값에 의한 자동차 이동 로직을 서비스로 이동
- 입력 스트링으로 자동차 객체를 생성하는 로직을 서비스로 이동
- 우승자 선정 로직을 서비스로 이동
