+++ src/main/kotlin/racingcar/Application.kt
@@ -1,12 +1,14 @@
 package racingcar
 
 import racingcar.controller.RacingGameController
-import racingcar.validator.InputValidator
+import racingcar.service.InputValidator
+import racingcar.service.RacingGameService
 import racingcar.view.RacingGameView
 
 fun main() {
     val racingGameView = RacingGameView()
     val inputValidator = InputValidator()
-    val racingGameController = RacingGameController(racingGameView, inputValidator)
+    val racingGameService = RacingGameService()
+    val racingGameController = RacingGameController(racingGameView, inputValidator, racingGameService)
     racingGameController.gameStart()
 }
