+++ src/main/kotlin/racingcar/service/InputValidator.kt
@@ -1,4 +1,4 @@
-package racingcar.validator
+package racingcar.service
 
 import racingcar.resources.Messages
